list選択画面ができた
