+++ package.json
@@ -15,9 +15,11 @@
     "@material-ui/icons": "^4.11.3",
     "@mui/lab": "^5.0.0-alpha.80",
     "@mui/material": "^5.6.3",
+    "axios": "^0.27.2",
     "next": "12.1.5",
     "react": "18.0.0",
     "react-dom": "18.0.0",
+    "react-router-dom": "^6.3.0",
     "yarn": "^1.22.18"
   },
   "devDependencies": {
